Fix for the arm time delays and general run improvements
Fixed the arm delays and got a 33 second run.
Potential issue with the RPS searching for the position and not finding it

+++ FEHRobot/Commands/ChangeArmPosition.h
@@ -22,8 +22,8 @@ public:
 private:
     Robot *robot;
 
-    int startTime;
-    int timeToWait;
+    float startTime;
+    float timeToWait;
     ArmPosition armPosition;
 
     StatusCode constructor(Robot *robot_, ArmPosition armPosition_, float sec);
